Add single RPM support to `spincoater_with_pc` and other improvements
- Update dependencies
- Add rounding to plate to/from motor RPM conversions
- Remove dev socket feature. We can't easily make host_tui generic over
  IO because `SerialPort` allows IO from immutable borrows while
  `std::io` doesn't.
- Make it more clear which commands will work while the MCU is running.

+++ rust/sc_messages/src/lib.rs
@@ -7,7 +7,7 @@ pub mod pwm;
 pub mod vacuum_pump;
 
 /// The number of motor revolutions per [`PLATE_REVOLUTIONS`] plate revolutions.
-pub const MOTOR_REVOLUTIONS: u32 = 72;
+pub const MOTOR_REVOLUTIONS: u16 = 72;
 
 /// The number of plate revolutions per [`MOTOR_REVOLUTIONS`] motor revolutions.
-pub const PLATE_REVOLUTIONS: u32 = 30;
+pub const PLATE_REVOLUTIONS: u16 = 30;
